+++ types-augment/lookup/types-substrate.ts
@@ -1,13 +1,13 @@
 // Auto-generated via `yarn polkadot-types-from-defs`, do not edit
 /* eslint-disable */
 
-declare module 'https://deno.land/x/polkadot@0.0.0-9/types/lookup.ts' {
-  import type { Data } from 'https://deno.land/x/polkadot@0.0.0-9/types/mod.ts';
-  import type { BTreeMap, Bytes, Compact, Enum, Null, Option, Result, Set, Struct, Text, U256, U8aFixed, Vec, WrapperKeepOpaque, bool, u128, u16, u32, u64, u8 } from 'https://deno.land/x/polkadot@0.0.0-9/types-codec/mod.ts';
-  import type { ITuple } from 'https://deno.land/x/polkadot@0.0.0-9/types-codec/types/index.ts';
-  import type { Vote } from 'https://deno.land/x/polkadot@0.0.0-9/types/interfaces/elections/index.ts';
-  import type { AccountId32, Call, H256, MultiAddress, PerU16, Perbill, Percent, Perquintill } from 'https://deno.land/x/polkadot@0.0.0-9/types/interfaces/runtime/index.ts';
-  import type { Event } from 'https://deno.land/x/polkadot@0.0.0-9/types/interfaces/system/index.ts';
+declare module 'https://deno.land/x/polkadot/types/lookup.ts' {
+  import type { Data } from 'https://deno.land/x/polkadot/types/mod.ts';
+  import type { BTreeMap, Bytes, Compact, Enum, Null, Option, Result, Set, Struct, Text, U256, U8aFixed, Vec, WrapperKeepOpaque, bool, u128, u16, u32, u64, u8 } from 'https://deno.land/x/polkadot/types-codec/mod.ts';
+  import type { ITuple } from 'https://deno.land/x/polkadot/types-codec/types/index.ts';
+  import type { Vote } from 'https://deno.land/x/polkadot/types/interfaces/elections/index.ts';
+  import type { AccountId32, Call, H256, MultiAddress, PerU16, Perbill, Percent, Perquintill } from 'https://deno.land/x/polkadot/types/interfaces/runtime/index.ts';
+  import type { Event } from 'https://deno.land/x/polkadot/types/interfaces/system/index.ts';
 
   /** @name FrameSystemAccountInfo (3) */
   export interface FrameSystemAccountInfo extends Struct {