+++ types/interfaces/eth/types.ts
@@ -1,9 +1,9 @@
 // Auto-generated via `yarn polkadot-types-from-defs`, do not edit
 /* eslint-disable */
 
-import type { GenericEthereumAccountId, GenericEthereumLookupSource } from 'https://deno.land/x/polkadot@0.0.0-9/types/mod.ts';
-import type { Bytes, Enum, Option, Struct, U256, U64, U8aFixed, Vec, bool, u32, u64 } from 'https://deno.land/x/polkadot@0.0.0-9/types-codec/mod.ts';
-import type { BlockNumber, H160, H2048, H256, H64 } from 'https://deno.land/x/polkadot@0.0.0-9/types/interfaces/runtime/index.ts';
+import type { GenericEthereumAccountId, GenericEthereumLookupSource } from 'https://deno.land/x/polkadot/types/mod.ts';
+import type { Bytes, Enum, Option, Struct, U256, U64, U8aFixed, Vec, bool, f64, u32, u64 } from 'https://deno.land/x/polkadot/types-codec/mod.ts';
+import type { BlockNumber, H160, H2048, H256, H64 } from 'https://deno.land/x/polkadot/types/interfaces/runtime/index.ts';
 
 /** @name BlockV0 */
 export interface BlockV0 extends Struct {
@@ -112,6 +112,14 @@ export interface EthereumLookupSource extends GenericEthereumLookupSource {}
 
 /** @name EthereumSignature */
 export interface EthereumSignature extends U8aFixed {}
+
+/** @name EthFeeHistory */
+export interface EthFeeHistory extends Struct {
+  readonly oldestBlock: U256;
+  readonly baseFeePerGas: Vec<U256>;
+  readonly gasUsedRatio: Vec<f64>;
+  readonly reward: Option<Vec<Vec<U256>>>;
+}
 
 /** @name EthFilter */
 export interface EthFilter extends Struct {